Assignment 2 update
Added bootstrap and jquery functionality to login, register and upload service.

Redone some of the CSS in the upload section.
Updated the upload section functionality, where you now can better preview the image to be uploaded, you can drag and drop, and remove the image from being uploaded. 

Moved HTML echos from the register endpoint to the RegisterService.

LoginService created, needs to be done for the login
HTML echos as in RegisterService.

Added hover animation for forms.

Added bootstrap alerts for regexChecks on the server side.